Add files via upload
The goal of the project is to recommend books to users based on their preferences. To accomplish this, We have applied Some algorithms:

Before the algorithm we have cleaned all data (performed EDA) in detail so that we got a well cleaned data.

Popularity Based Filtering: This algorithm recommends popular items to users. The idea is that if many people like a particular item, it's likely that the user will like it too. This algorithm is simple and easy to implement but may not be very personalized.

Collaborative Filtering Based Recommender System: Collaborative filtering is a technique that recommends items to users based on their similarity to other users. The algorithm looks for patterns in the data to identify similar users and recommend items that those users have liked in the past. This algorithm can be more personalized than popularity-based filtering, but it requires a large amount of data to be effective.

Optimization of SVD algorithm: Singular Value Decomposition (SVD) is a matrix factorization technique that can be used for collaborative filtering. It decomposes the user-item interaction matrix into lower-dimensional matrices that represent the latent factors that influence user preferences. Optimizing the SVD algorithm involves finding the optimal values for the hyperparameters that control the decomposition. This can improve the accuracy of the recommendations.

Finding the Nearest Neighbors: This algorithm uses the k-nearest neighbors (k-NN) technique to find the items that are most similar to the user's preferred items. It calculates the similarity between items based on their attributes (e.g., genre, author, publication date). The algorithm then recommends items that are similar to the user's preferred items.

To summarize, the project uses a combination of popularity-based filtering, collaborative filtering, optimization of the SVD algorithm, and finding the nearest neighbors to recommend books to users based on their preferences. These algorithms can provide a personalized and accurate recommendation system that can help users discover new books that they may enjoy.